+++ packages/hiccup-html/src/index.ts
@@ -1,8 +1,10 @@
 export * from "./api";
 export * from "./def";
 
-export * from "./anchor";
 export * from "./blocks";
-export * from "./input";
+export * from "./forms";
+export * from "./head";
+export * from "./inline";
+export * from "./lists";
 export * from "./media";
 export * from "./sections";