+++ packages/hiccup-html/src/lists.ts
@@ -1,0 +1,24 @@
+import { Attribs, AttribVal } from "./api";
+import { defElement, defElements } from "./def";
+
+export interface OrderedListAttribs extends Attribs {
+    reversed: AttribVal<boolean>;
+    start: AttribVal<number>;
+    type: AttribVal<"1" | "a" | "A" | "i" | "I">;
+}
+
+export interface ListItemAttribs extends Attribs {
+    value: AttribVal<number>;
+}
+
+export const ol = defElement<Partial<OrderedListAttribs>>("ol");
+
+export const li = defElement<Partial<ListItemAttribs>>("li");
+
+export const [ul, dl, dt, dd, datalist] = defElements([
+    "ul",
+    "dl",
+    "dt",
+    "dd",
+    "datalist",
+]);